Add Database _ 데이터베이스 기반 생성 # 세부사항 1. 데이터베이스 초기화 및 기본적인 쿼리 실행 함수 생성 - 쿼리를 실행할 때마다 멤버 구조체에 데이터가 바뀜. 해당 변수를 통해 각자의 구조체로 매핑시켜 줘야 함 2. 자료형 크기 경고 해결 - int형을 전부 int64_t형으로 변환. 넘어오는 자료는 모두 8바이트 포인터를 가짐
# 해야할 일
1. db 테이블 생성후 서버 로직과 연동 ( 기본적으로 id가 존재하는지 여부 - 로그인 과정)

+++ KaKaoProgram/Inc/Stream.h
@@ -2,31 +2,31 @@
 #include "ServLibrary.h"
 
 #define SHORT_SIZE sizeof(short)
-#define INT_SIZE sizeof(int32_t)
+#define INT_SIZE sizeof(int64_t)
 
 class OutputStream
 {
 public:
 
-	void Write(const char* pData, const int dataSize)
+	void Write(const char* pData, const int64_t dataSize)
 	{
-		int next = m_head + dataSize;
+		int64_t next = m_head + dataSize;
 		const char* copyData = static_cast<const char*>(pData);
 		if (next > m_capacity)
 		{
-			ReallocBuffer(std::max<int>(m_capacity * 2, next));
+			ReallocBuffer(std::max<int64_t>(m_capacity * 2, next));
 		}
 
 		memcpy(&m_buffer[m_head], &copyData[0], dataSize);
 		m_head += dataSize;
 	}
 	void Write(const std::string inString);
-	void Write(const short data, const int32_t dataSize = SHORT_SIZE) { Write((const char*)data, dataSize); }
-	void Write(const int32_t data, const int32_t dataSize = INT_SIZE) { Write((const char*)data, dataSize); }
-	void ReallocBuffer(int32_t BufferCapacity);
+	void Write(const short data, const int64_t dataSize = SHORT_SIZE) { Write((const char*)data, dataSize); }
+	void Write(const int64_t data, const int64_t dataSize = INT_SIZE) { Write((const char*)data, dataSize); }
+	void ReallocBuffer(int64_t BufferCapacity);
 private:
-	int m_head;
-	int m_capacity;
+	int64_t m_head;
+	int64_t m_capacity;
 	char* m_buffer;
 public:
 	/// Capcaity는 최대한 패킷크기보다 훨신 큰 크기로 잡아준다.
@@ -36,20 +36,20 @@ public:
 class InputStream
 {
 private:
-	void ReallocBuffer(int32_t BufferCapacity);
+	void ReallocBuffer(int64_t BufferCapacity);
 public:
 
-	void Read(char* pData, const int32_t dataSize)
+	void Read(char* pData, const int64_t dataSize)
 	{
-		int next = m_head + dataSize;
+		int64_t next = m_head + dataSize;
 		const char* copyData = reinterpret_cast<const char*>(pData);
 
 		memcpy((void*)&copyData[0], &m_buffer[m_head], dataSize);
 		m_head = next;
 	}
 	void Read(std::string& inString);
-	void Read(short data, const int32_t dataSize = SHORT_SIZE) { Read((char*)data, dataSize); }
-	void Read(int32_t data, const int32_t dataSize = INT_SIZE) { Read((char*)data, dataSize); }
+	void Read(short data, const int64_t dataSize = SHORT_SIZE) { Read((char*)data, dataSize); }
+	void Read(int64_t data, const int64_t dataSize = INT_SIZE) { Read((char*)data, dataSize); }
 
 	char* GetBuffer()const { return m_buffer; }
 public:
@@ -64,10 +64,10 @@ public:
 		return *this;
 	}
 public:
-	InputStream(const int32_t Capacity, char* buffer);
+	InputStream(const int64_t Capacity, char* buffer);
 	~InputStream();
 private:
-	int32_t m_head;
-	int32_t m_capacity;
+	int64_t m_head;
+	int64_t m_capacity;
 	char* m_buffer;
 };
