+++ KaKaoProgram/Src/Stream.cpp
@@ -20,7 +20,7 @@ void OutputStream::Write(const std::string inString)
 	}
 }
 
-void OutputStream::ReallocBuffer(int32_t BufferCapacity)
+void OutputStream::ReallocBuffer(int64_t BufferCapacity)
 {
 	char* copyBuffer = new char[BufferCapacity];
 	if (m_buffer == nullptr)
@@ -35,7 +35,7 @@ void OutputStream::ReallocBuffer(int32_t BufferCapacity)
 	}
 }
 
-void InputStream::ReallocBuffer(int32_t BufferCapacity)
+void InputStream::ReallocBuffer(int64_t BufferCapacity)
 {
 	char* copyBuffer = new char[BufferCapacity];
 	if (m_buffer == nullptr)
@@ -52,7 +52,7 @@ void InputStream::ReallocBuffer(int32_t BufferCapacity)
 
 void InputStream::Read(std::string& inString)
 {
-	int32_t dataSize = 0;
+	int64_t dataSize = 0;
 	Read((char*)&dataSize, inString.size());
 	inString.resize(dataSize);
 	for (auto& elem : inString)
@@ -61,7 +61,7 @@ void InputStream::Read(std::string& inString)
 	}
 }
 
-InputStream::InputStream(const int32_t Capacity, char* buffer) :m_head(0), m_capacity(Capacity), m_buffer(buffer)
+InputStream::InputStream(const int64_t Capacity, char* buffer) :m_head(0), m_capacity(Capacity), m_buffer(buffer)
 {
 	ReallocBuffer(Capacity);
 }
